Update App to build Exchange from py script, as well as Read from .cs…

+++ include/Stock.h
@@ -16,7 +16,7 @@ private:
     float dayChange;
     float percentChange; 
 public:
-    Stock(std::string tick, int val);
+    Stock(std::string tick, float val);
     ~Stock();
 
     float getValue();
